월급 달력에서 n.5시간도 월급 제대로 나오게 고쳤어용

+++ app/src/main/java/com/ssuandroid/my_parttime/CalendarFragment.java
@@ -162,7 +162,7 @@ public class CalendarFragment extends Fragment {
                                     String dateString = timestampToDate.toString();
                                     if (timestampToDate.after(end) && timestampToDate.before(start)) {
                                         String branchName = (String) document.get("branchName");
-                                        totalWage += (((Number) document.get("wage")).intValue() * ((Number) document.get("workTime")).intValue());
+                                        totalWage += (((Number) document.get("wage")).intValue() * ((Number) document.get("workTime")).doubleValue());
                                         String day = null;
                                         SimpleDateFormat dateFormat = new SimpleDateFormat("EEE MMM dd HH:mm:ss zzz yyyy", Locale.US);
                                         try {
